データ型の修正

+++ boundary_condition.py
@@ -119,7 +119,7 @@ class DyeBoundaryCondition(BoundaryCondition):
 def create_boundary_condition1(resolution, no_dye=False):
     # 1: 壁, 2: 流入部, 3: 流出部
     bc = np.zeros((2 * resolution, resolution, 2), dtype=np.float32)
-    bc_mask = np.zeros((2 * resolution, resolution), dtype=np.uint32)
+    bc_mask = np.zeros((2 * resolution, resolution), dtype=np.uint8)
     bc_dye = np.zeros((2 * resolution, resolution, 3), dtype=np.float32)
 
     # 流入部の設定
@@ -154,7 +154,7 @@ def create_boundary_condition1(resolution, no_dye=False):
 
 def create_boundary_condition2(resolution, no_dye=False):
     bc = np.zeros((2 * resolution, resolution, 2), dtype=np.float32)
-    bc_mask = np.zeros((2 * resolution, resolution), dtype=np.uint32)
+    bc_mask = np.zeros((2 * resolution, resolution), dtype=np.uint8)
     bc_dye = np.zeros((2 * resolution, resolution, 3), dtype=np.float32)
 
     # 流入部の設定
@@ -212,7 +212,7 @@ def create_boundary_condition2(resolution, no_dye=False):
 
 def create_boundary_condition3(resolution, no_dye=False):
     bc = np.zeros((2 * resolution, resolution, 2), dtype=np.float32)
-    bc_mask = np.zeros((2 * resolution, resolution), dtype=np.uint32)
+    bc_mask = np.zeros((2 * resolution, resolution), dtype=np.uint8)
     bc_dye = np.zeros((2 * resolution, resolution, 3), dtype=np.float32)
 
     # 流入部の設定
@@ -252,7 +252,7 @@ def create_boundary_condition3(resolution, no_dye=False):
 
 def create_boundary_condition4(resolution, no_dye=False):
     bc = np.zeros((2 * resolution, resolution, 2), dtype=np.float32)
-    bc_mask = np.zeros((2 * resolution, resolution), dtype=np.uint32)
+    bc_mask = np.zeros((2 * resolution, resolution), dtype=np.uint8)
     bc_dye = np.zeros((2 * resolution, resolution, 3), dtype=np.float32)
 
     # 壁の設定
